Developed software for testing and modified model to take such input. Everytime a new file is generated it must be updated in the model though.
The python app is rest and uses only a POST in /generateScenario following this json:

{
 "nombre":"Base1",
 "d":4,
 "n":4,
 "maxd":1,
 "mind":1,
 "inicio":1,
 "p1":1,
 "p2":0

}

Optionally you can add the params "score","reviews","costoVida","costoTrabajo" that are the name of files for uploading data, if the value is "Random" it adds random integer numbers from 1 to 100000. And if it is not present, it adds 1s to all possible values.